chore(webpanel/pages/settings): moved audit log row to separate component
